Nested LayerProperties class within GetPrintHandler Made GetPrintHandler read the service endpoint url from property instead of using the hardcoded value Removed unnecessary stuff from webapp-print/web.xml
