+++ control-base/src/main/java/fi/nls/oskari/control/data/GetPrintHandler.java
@@ -13,7 +13,6 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.json.JSONObject;
 import org.oskari.print.PrintFormat;
 import org.oskari.print.PrintLayer;
 import org.oskari.print.PrintRequest;
@@ -35,8 +34,8 @@ import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.service.ProxyService;
 import fi.nls.oskari.util.ConversionHelper;
 import fi.nls.oskari.util.IOHelper;
+import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ServiceFactory;
-import fi.nls.oskari.wmts.domain.TileMatrix;
 
 @OskariActionRoute("GetPrint")
 public class GetPrintHandler extends ActionHandler {
@@ -53,8 +52,8 @@ public class GetPrintHandler extends ActionHandler {
     private static final String PARM_FORMAT = "format";
     private static final String PARM_SRSNAME = "srs";
 
-    private static final int DPI = 72;
     private static final double MM_PER_INCH = 25.4;
+    public static final double OGC_DPI = MM_PER_INCH / 0.28;
 
     private static final int A4W = 210;
     private static final int A4H = 297;
@@ -65,12 +64,12 @@ public class GetPrintHandler extends ActionHandler {
     private static final int MARGIN_WIDTH = 10 * 2;
     private static final int MARGIN_HEIGHT = 15 * 2;
 
-
-    private static final String PRINT_URL = "http://localhost:8080/print/out/";
+    private String servicePrintURL;
     private PermissionHelper permissionHelper;
 
 
     public void init() {
+        servicePrintURL = PropertyUtil.get("service.print.url");
         permissionHelper = new PermissionHelper(
                 ServiceFactory.getMapLayerService(), 
                 ServiceFactory.getPermissionsService());
@@ -84,7 +83,7 @@ public class GetPrintHandler extends ActionHandler {
             PrintRequest request = createPrintRequest(params);
             byte[] json = OBJECT_MAPPER.writeValueAsBytes(request);
 
-            HttpURLConnection con = IOHelper.getConnection(PRINT_URL);
+            HttpURLConnection con = IOHelper.getConnection(servicePrintURL);
             copyHeaders(con, params.getRequest());
             postRequest(con, json);
             passResponseThrough(con, params.getResponse());
@@ -95,7 +94,7 @@ public class GetPrintHandler extends ActionHandler {
 
 
     public static int mmToPx(int mm) {
-        return (int) Math.round((double) (DPI * mm) / MM_PER_INCH);
+        return (int) Math.round((OGC_DPI * mm) / MM_PER_INCH);
     }
 
 
@@ -206,7 +205,9 @@ public class GetPrintHandler extends ActionHandler {
         for (LayerProperties requestedLayer : requestedLayers) {
             OskariLayer oskariLayer = permissionHelper.getLayer(requestedLayer.getId(), user);
             PrintLayer printLayer = createPrintLayer(oskariLayer, requestedLayer);
-            printLayers.add(printLayer);
+            if (printLayer != null) {
+                printLayers.add(printLayer);
+            }
         }
 
         return printLayers;
@@ -214,6 +215,12 @@ public class GetPrintHandler extends ActionHandler {
 
 
     private PrintLayer createPrintLayer(OskariLayer oskariLayer, LayerProperties requestedLayer) {
+        int opacity = getOpacity(requestedLayer.getOpacity(), oskariLayer.getOpacity());
+        if (opacity <= 0) {
+            // Ignore fully transparent layers
+            return null;
+        }
+
         PrintLayer layer = new PrintLayer();
         layer.setId(oskariLayer.getExternalId());
         layer.setType(oskariLayer.getType());
@@ -222,11 +229,11 @@ public class GetPrintHandler extends ActionHandler {
         layer.setName(oskariLayer.getName());
         layer.setUsername(oskariLayer.getUsername());
         layer.setPassword(oskariLayer.getPassword());
-        layer.setOpacity(getOpacity(requestedLayer.getOpacity(), oskariLayer.getOpacity()));
+        layer.setOpacity(opacity);
         return layer;
     }
-    
-    
+
+
     private int getOpacity(int requestedOpacity, Integer layersDefaultOpacity) {
         int opacity;
         if (requestedOpacity != LayerProperties.NULL) {
@@ -236,14 +243,11 @@ public class GetPrintHandler extends ActionHandler {
             // Otherwise use layers default opacity
             opacity = layersDefaultOpacity;
         } else {
-            // If that's missing or negative use 100 (full opacity)
+            // If that's missing just use 100 (full opacity)
             opacity = 100;
         }
 
-        // Scale opacity to [0, 100]
-        if (opacity < 0) {
-            opacity = 0;
-        } else if (opacity > 100) {
+        if (opacity > 100) {
             opacity = 100;
         }
 
@@ -251,17 +255,16 @@ public class GetPrintHandler extends ActionHandler {
     }
 
 
-    // TODO: Replace parse logic with a simple POJO implementation
     private void setTiles(ActionParameters params, List<PrintLayer> layers)
             throws ActionException {
         try {
             String tilesJson = params.getHttpParam("PARM_TILES");
-            
+
             JsonNode root = OBJECT_MAPPER.readTree(tilesJson);
             if (!root.isArray()) {
                 throw new ActionParamsException("Tiles not array!");
             }
-            
+
             Iterator<String> it = root.fieldNames();
             while (it.hasNext()) {
                 String layerId = it.next();
@@ -269,24 +272,24 @@ public class GetPrintHandler extends ActionHandler {
                 if (printLayer == null) {
                     continue;
                 }
-                
+
                 JsonNode layerNode = root.get(layerId);
                 int n = layerNode.size();
                 if (n != 1) {
                     throw new ActionParamsException("Bad tiles! Size not 1");
                 }
-                
+
                 JsonNode tilesNode = layerNode.get(0);
                 if (!tilesNode.isArray()) {
                     throw new ActionParamsException("Bad tiles! Not array");
                 }
-                
+
                 int m = tilesNode.size();
                 Tile[] tiles = new Tile[m];
-                
+
                 for (int j = 0; j < m; j++) {
                     JsonNode tileNode = tilesNode.get(j);
-                    
+
                     JsonNode bboxNode = tileNode.get("bbox");
                     if (bboxNode == null 
                             || !bboxNode.isArray() 
@@ -294,22 +297,22 @@ public class GetPrintHandler extends ActionHandler {
                         throw new ActionParamsException("Bad tiles! "
                                 + "Missing 'bbox' or not array or not size 4");
                     }
-                    
+
                     JsonNode urlNode = tileNode.get("url");
                     if (urlNode == null || !urlNode.isTextual()) {
                         throw new ActionParamsException("Bad tiles! "
                                 + "Missing 'url' or not text");
                     }
-                    
+
                     double[] bbox = new double[4]; 
                     for (int k = 0; k < 4; k++) {
                         bbox[k] = bboxNode.get(k).asDouble();
                     }
                     String url = urlNode.asText();
-                    
+
                     tiles[j] = new Tile(bbox, url);
                 }
-                
+
                 printLayer.setTiles(tiles);
             }
         } catch (IOException e) {
@@ -364,6 +367,70 @@ public class GetPrintHandler extends ActionHandler {
                 OutputStream out = response.getOutputStream()) {
             IOHelper.copy(in, out);
         }
+    }
+
+
+    public static class LayerProperties {
+
+        private static final int NULL = Integer.MIN_VALUE;
+
+        private final String id;
+        private final int opacity;
+        private final String style;
+
+        public LayerProperties(String id, int opacity, String style) {
+            this.id = id;
+            this.opacity = opacity;
+            this.style = style;
+        }
+
+        public String getId() {
+            return id;
+        }
+
+        public int getOpacity() {
+            return opacity;
+        }
+
+        public String getStyle() {
+            return style;
+        }
+
+        public static LayerProperties parse(String[] layerParam) 
+                throws IllegalArgumentException {
+            if (layerParam == null || layerParam.length == 0) {
+                throw new IllegalArgumentException(
+                        "layerParam must be non-null with positive length!");
+            }
+
+            String id = layerParam[0];
+            int opacity = layerParam.length > 1 ?
+                    ConversionHelper.getInt(layerParam[1], NULL) : NULL;
+                    String style = layerParam.length > 2 ? layerParam[2] : null;
+
+                    return new LayerProperties(id, opacity, style);
+        }
+
+
+        public static LayerProperties[] parse(String mapLayers) 
+                throws IllegalArgumentException {
+            final String[] layers = mapLayers.split(",");
+            final int n = layers.length;
+
+            final LayerProperties[] layerProperties = new LayerProperties[n];
+            for (int i = 0; i < n; i++) {
+                String[] layerParam = layers[i].split(" ");
+                layerProperties[i] = parse(layerParam);
+            }
+
+            return layerProperties;
+        }
+
+
+        public static int getDefaultOpacity() {
+            return NULL;
+        }
+
     }
 
 }
