+++ webapp-print/src/main/webapp/WEB-INF/web.xml
@@ -4,29 +4,6 @@
          xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
          version="3.0">
 
-	<!-- 
-	<context-param>
-	    <param-name>org.eclipse.jetty.servlet.SessionCookie</param-name>
-	    <param-value>WFSSESSIONID</param-value>
-	</context-param>
-	
-	<context-param>
-	    <param-name>org.eclipse.jetty.servlet.SessionIdPathParameterName</param-name>
-	    <param-value>wfssessionid</param-value>
-	</context-param>
-	
-	<filter>
-        <filter-name>cross-origin</filter-name>
-        <filter-class>org.eclipse.jetty.servlets.CrossOriginFilter</filter-class>
-        <async-supported>true</async-supported>
-    </filter>
-    
-    <filter-mapping>
-        <filter-name>cross-origin</filter-name>
-        <url-pattern>/print/*</url-pattern>
-    </filter-mapping>
-     -->
-
 	<servlet>
         <servlet-name>print</servlet-name>
         <servlet-class>org.oskari.print.PrintServlet</servlet-class>
